Change some clocks to support sub-tick precision using a new class, ClockTicks. Using that higher precision, interpolate at a higher granularity. Further, truncate the target interpolation time to vertical blank boundaries to avoid producing temporal artifacts. Fix issues caused by interpolation calculations being handled differently in multiple places (and fix cases where smoothratios were being thrown away only to be redone without checking all proper conditions). Ensure ClockTicks changes do not break other targets (EKenBuild, VoidSW), but note any interpolation there is not similarly updated.
git-svn-id: https://svn.eduke32.com/eduke32@8050 1a8010ca-5511-0410-912e-c29ae57300e0

# Conflicts:
#	source/build/src/build.cpp
#	source/build/src/sdlayer.cpp

+++ source/duke3d/src/actors.cpp
@@ -326,13 +326,13 @@ next_sprite:
 // <fromunderp>: below->above change?
 static int32_t Proj_MaybeDoTransport(int32_t spriteNum, uspriteptr_t const pSEffector, int32_t fromunderp, int32_t daz)
 {
-    if ((totalclock & UINT8_MAX) == actor[spriteNum].lasttransport)
+    if (((int32_t) totalclock & UINT8_MAX) == actor[spriteNum].lasttransport)
         return 0;
 
     auto const pSprite = &sprite[spriteNum];
     auto const otherse = (uspriteptr_t)&sprite[pSEffector->owner];
 
-    actor[spriteNum].lasttransport = (totalclock & UINT8_MAX);
+    actor[spriteNum].lasttransport = ((int32_t) totalclock & UINT8_MAX);
 
     pSprite->x += (otherse->x - pSEffector->x);
     pSprite->y += (otherse->y - pSEffector->y);
@@ -3623,7 +3623,7 @@ ACTOR_STATIC void G_MoveTransports(void)
                 case STAT_FALLER:
                 case STAT_DUMMYPLAYER:
                 {
-                    if ((totalclock & UINT8_MAX) != actor[sectSprite].lasttransport)
+                    if (((int32_t) totalclock & UINT8_MAX) != actor[sectSprite].lasttransport)
                     {
                         int const zvel    = sprite[sectSprite].zvel;
                         int const absZvel = klabs(zvel);
@@ -3699,7 +3699,7 @@ ACTOR_STATIC void G_MoveTransports(void)
                                             A_SetSprite(newSprite, CLIPMASK0);
                                         }
 #endif
-                                        actor[sectSprite].lasttransport = (totalclock & UINT8_MAX);
+                                        actor[sectSprite].lasttransport = ((int32_t) totalclock & UINT8_MAX);
 
                                         sprite[sectSprite].x += sprite[OW(spriteNum)].x - SX(spriteNum);
                                         sprite[sectSprite].y += sprite[OW(spriteNum)].y - SY(spriteNum);
