+++ source/duke3d/src/anim.cpp
@@ -512,7 +512,7 @@ int32_t Anim_Play(const char *fn)
             g_restorePalette = 0;
         }
 
-        frametime = totalclock;
+        frametime = (int32_t) totalclock;
 
         videoClearScreen(0);
