+++ source/duke3d/src/sector.cpp
@@ -445,7 +445,7 @@ void G_AnimateCamSprite(int smoothRatio)
 #endif
         }
 
-        T1(spriteNum) = totalclock;
+        T1(spriteNum) = (int32_t) totalclock;
     }
 }
 
@@ -1210,10 +1210,10 @@ int P_ActivateSwitch(int playerNum, int wallOrSprite, int switchType)
 
     if (switchType == SWITCH_SPRITE) // A wall sprite
     {
-        if (actor[wallOrSprite].lasttransport == (totalclock & UINT8_MAX))
+        if (actor[wallOrSprite].lasttransport == ((int32_t) totalclock & UINT8_MAX))
             return 0;
 
-        actor[wallOrSprite].lasttransport = (totalclock & UINT8_MAX);
+        actor[wallOrSprite].lasttransport = ((int32_t) totalclock & UINT8_MAX);
 
         if (sprite[wallOrSprite].lotag == 0)
             return 0;
